use nuxt-workers
