+++ app/composables/useSearch.ts
@@ -1,13 +1,8 @@
 import type { SearchOptions, SearchResult } from 'comark-content'
-import type { SearchWorkerPayload, SearchWorkerResponse } from '../types/search-worker'
 
 type SearchStatus = 'idle' | 'loading' | 'ready' | 'error'
 
 const status = ref<SearchStatus>('idle')
-
-let worker: Worker | undefined
-let nextId = 0
-const pending = new Map<number, { resolve: (results: SearchResult[]) => void, reject: (error: Error) => void }>()
 
 /**
  * Hydration logging switch: `?debug=search`
@@ -15,50 +10,6 @@ const pending = new Map<number, { resolve: (results: SearchResult[]) => void, re
 function searchDebug(): boolean {
   if (!import.meta.client) return false
   return new URLSearchParams(location.search).get('debug') === 'search'
-}
-
-function getWorker(): Worker {
-  if (worker) return worker
-
-  worker = new Worker(new URL('../workers/search.worker.ts', import.meta.url), { type: 'module' })
-
-  worker.onmessage = (event: MessageEvent<SearchWorkerResponse>) => {
-    const message = event.data
-    if (message.type === 'status') {
-      status.value = message.value
-      if (searchDebug()) console.info(`[search] status -> ${message.value}`)
-      return
-    }
-    const settle = pending.get(message.id)
-    if (!settle) return
-    pending.delete(message.id)
-    if (message.type === 'result') settle.resolve(message.results)
-    else {
-      if (searchDebug()) console.error(`[search] request ${message.id} failed:`, message.message)
-      settle.reject(new Error(message.message))
-    }
-  }
-
-  worker.onerror = () => {
-    status.value = 'error'
-    for (const { reject } of pending.values()) reject(new Error('[search] the search worker failed to load'))
-    pending.clear()
-  }
-
-  return worker
-}
-
-function request(message: SearchWorkerPayload): Promise<SearchResult[]> {
-  const id = ++nextId
-  return new Promise<SearchResult[]>((resolve, reject) => {
-    pending.set(id, { resolve, reject })
-    try {
-      getWorker().postMessage({ ...message, id })
-    } catch (error) {
-      pending.delete(id)
-      reject(error instanceof Error ? error : new Error(String(error)))
-    }
-  })
 }
 
 /**
@@ -73,10 +24,12 @@ export function useSearch() {
   )
 
   /**
-   * Load the database ahead of the first keystroke. No-op once loading or ready; retries after a
-   * failure — the worker holds that guard, since this side's `status` lags a message behind.
+   * Load the database.
+   * No-op once loading or ready; retries after a failure.
    */
   async function warmup(): Promise<void> {
+    if (status.value === 'loading' || status.value === 'ready') return
+    status.value = 'loading'
     try {
       if (!headSha.value && !import.meta.dev) {
         throw new Error('[search] /api/content/head returned no commit pin')
@@ -88,7 +41,8 @@ export function useSearch() {
       const debug = searchDebug()
       if (debug) console.info(`[search] warmup from ${apiBase} (head ${headSha.value ?? 'unpinned'})`)
 
-      await request({ type: 'warmup', apiBase, origin: location.origin, debug })
+      await warmupSearch(apiBase, location.origin, debug)
+      status.value = 'ready'
     } catch (error) {
       status.value = 'error'
       console.error('[search] could not load the search database', error)
@@ -100,7 +54,7 @@ export function useSearch() {
   }
 
   async function search(query: string, opts?: SearchOptions): Promise<SearchResult[]> {
-    return request({ type: 'search', query, opts })
+    return searchContent(query, opts)
   }
 
   return { search, status: readonly(status), warmup }
